requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
SQLAlchemy==2.0.36
pydantic==2.8.2
python-dotenv==1.0.1
psycopg==3.1.18
Jinja2==3.1.4
pydantic[email]
passlib==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.9
